Set currency of selected tokens by roll expression
Macro allows you to set the currency of the selected tokens using rollable expressions. This works well with the loot sheet to provide random currency to each individual NPC. It also gives a selector that allows the user to change how the new values are applied to each sheet. The default is "Replace Value", but there options to Add, Subtract, Multiply, and Divide as well.